changed magnetic forcing to be an emf inside the curl.  NOT TESTED
:254:851297ad-b6b1-409f-8f61-2fb168c4d1e3:2010-07-15T21:12:20.000000Z:

+++ src/include/LogTrace.h
@@ -7,18 +7,18 @@
 
 //TODO Add timing functionality to the other log levels
 #undef trace
-#define trace(format, ...) fprintf(procFile, "TRACE: %s %d - %d: ", __FILE__, __LINE__, clock()); fprintf(procFile, format, __VA_ARGS__); fflush(procFile);
+#define trace(format, ...) fprintf(procFile, "TRACE: %s %d - %f: ", __FILE__, __LINE__, (double)clock()); fprintf(procFile, format, __VA_ARGS__); fflush(procFile);
 
 #undef debug
-#define debug(format, ...) fprintf(procFile, "DEBUG: %s %d - %d: ", __FILE__, __LINE__, clock()); fprintf(procFile, format, __VA_ARGS__); fflush(procFile);
+#define debug(format, ...) fprintf(procFile, "DEBUG: %s %d - %f: ", __FILE__, __LINE__, (double)clock()); fprintf(procFile, format, __VA_ARGS__); fflush(procFile);
 
 #undef info
-#define info(format, ...) fprintf(procFile, "INFO:  %s %d - %d: ", __FILE__, __LINE__, clock()); fprintf(procFile, format, __VA_ARGS__); fflush(procFile);
+#define info(format, ...) fprintf(procFile, "INFO:  %s %d - %f: ", __FILE__, __LINE__, (double)clock()); fprintf(procFile, format, __VA_ARGS__); fflush(procFile);
 
 #undef warn
-#define warn(format, ...) fprintf(procFile, "WARN:  %s %d - %d: ", __FILE__, __LINE__, clock()); fprintf(procFile, format, __VA_ARGS__); fflush(procFile);
+#define warn(format, ...) fprintf(procFile, "WARN:  %s %d - %f: ", __FILE__, __LINE__, (double)clock()); fprintf(procFile, format, __VA_ARGS__); fflush(procFile);
 
 #undef error
-#define error(format, ...) fprintf(procFile, "ERROR: %s %d - %d: ", __FILE__, __LINE__, clock()); fprintf(procFile, format, __VA_ARGS__); fflush(procFile);
+#define error(format, ...) fprintf(procFile, "ERROR: %s %d - %f: ", __FILE__, __LINE__, (double)clock()); fprintf(procFile, format, __VA_ARGS__); fflush(procFile);
